requirements.txt
fastapi
uvicorn[standard]
pydantic>=2.7
google-cloud-vision
pytest
